Removed Main element as wrapper
for Play Area section and Control Panel Section as per discussion with
@jessegreenberg.
I’m comfortable with no wrapper. A section element needs a heading, so
nesting the PA and CP in a section doesn’t make sense to me. If a
wrapper is needed, it might have to be a div.